order-candidates REST-API - option to just lookup, but not create BPartners
* WIP - doesn't compile
* start to separate code:
  * bpartner-handling from the rest
  * lookup from creation/update
* IOLCandEffectiveValuesBL: change BPartnerId-methods' return type from int to BPartnerId
* BPartnerDAO change I_AD_User pacakge to the "simpler" org.compiere.model

https://github.com/metasfresh/metasfresh/issues/4756
