+++ de.metas.adempiere.adempiere/base/src/main/java-legacy/org/compiere/process/CommissionCalc.java
@@ -25,6 +25,7 @@ import java.util.GregorianCalendar;
 import java.util.List;
 
 import org.adempiere.ad.trx.api.ITrx;
+import org.compiere.model.I_AD_User;
 import org.compiere.model.MCommission;
 import org.compiere.model.MCommissionAmt;
 import org.compiere.model.MCommissionDetail;
@@ -35,7 +36,6 @@ import org.compiere.util.AdempiereUserError;
 import org.compiere.util.DB;
 import org.compiere.util.DisplayType;
 
-import de.metas.adempiere.model.I_AD_User;
 import de.metas.bpartner.service.IBPartnerDAO;
 import de.metas.currency.ICurrencyDAO;
 import de.metas.i18n.Language;
@@ -44,8 +44,8 @@ import de.metas.process.ProcessInfoParameter;
 import de.metas.util.Services;
 
 /**
- *	Commission Calculation	
- *	
+ *	Commission Calculation
+ *
  *  @author Jorg Janke
  *  @version $Id: CommissionCalc.java,v 1.3 2006/09/25 00:59:41 jjanke Exp $
  */
@@ -90,20 +90,20 @@ public class CommissionCalc extends JavaProcess
 		m_com = new MCommission (getCtx(), getRecord_ID(), get_TrxName());
 		if (m_com.get_ID() == 0)
 			throw new AdempiereUserError ("No Commission");
-			
-		//	Create Commission	
+
+		//	Create Commission
 		MCommissionRun comRun = new MCommissionRun (m_com);
 		setStartEndDate();
-		comRun.setStartDate(p_StartDate);		
+		comRun.setStartDate(p_StartDate);
 		//	01-Jan-2000 - 31-Jan-2001 - USD
 		SimpleDateFormat format = DisplayType.getDateFormat(DisplayType.Date);
-		String description = format.format(p_StartDate) 
+		String description = format.format(p_StartDate)
 			+ " - " + format.format(m_EndDate)
 			+ " - " + Services.get(ICurrencyDAO.class).getISO_Code(getCtx(), m_com.getC_Currency_ID());
 		comRun.setDescription(description);
 		if (!comRun.save())
 			throw new AdempiereSystemError ("Could not save Commission Run");
-		
+
 		MCommissionLine[] lines = m_com.getLines();
 		for (int i = 0; i < lines.length; i++)
 		{
@@ -254,15 +254,15 @@ public class CommissionCalc extends JavaProcess
 			comAmt.calculateCommission();
 			comAmt.save();
 		}	//	for all commission lines
-		
+
 	//	comRun.updateFromAmt();
 	//	comRun.save();
-		
+
 		//	Save Last Run
 		m_com.setDateLastRun (p_StartDate);
 		m_com.save();
-		
-		return "@C_CommissionRun_ID@ = " + comRun.getDocumentNo() 
+
+		return "@C_CommissionRun_ID@ = " + comRun.getDocumentNo()
 			+ " - " + comRun.getDescription();
 	}	//	doIt
 
@@ -284,9 +284,9 @@ public class CommissionCalc extends JavaProcess
 			p_StartDate = new Timestamp (cal.getTimeInMillis());
 			//
 			cal.add(Calendar.YEAR, 1);
-			cal.add(Calendar.DAY_OF_YEAR, -1); 
+			cal.add(Calendar.DAY_OF_YEAR, -1);
 			m_EndDate = new Timestamp (cal.getTimeInMillis());
-			
+
 		}
 		//	Quarterly
 		else if (MCommission.FREQUENCYTYPE_Quarterly.equals(m_com.getFrequencyType()))
@@ -304,7 +304,7 @@ public class CommissionCalc extends JavaProcess
 			p_StartDate = new Timestamp (cal.getTimeInMillis());
 			//
 			cal.add(Calendar.MONTH, 3);
-			cal.add(Calendar.DAY_OF_YEAR, -1); 
+			cal.add(Calendar.DAY_OF_YEAR, -1);
 			m_EndDate = new Timestamp (cal.getTimeInMillis());
 		}
 		//	Weekly
@@ -313,7 +313,7 @@ public class CommissionCalc extends JavaProcess
 			cal.set(Calendar.DAY_OF_WEEK, Calendar.SUNDAY);
 			p_StartDate = new Timestamp (cal.getTimeInMillis());
 			//
-			cal.add(Calendar.DAY_OF_YEAR, 7); 
+			cal.add(Calendar.DAY_OF_YEAR, 7);
 			m_EndDate = new Timestamp (cal.getTimeInMillis());
 		}
 		//	Monthly
@@ -323,11 +323,11 @@ public class CommissionCalc extends JavaProcess
 			p_StartDate = new Timestamp (cal.getTimeInMillis());
 			//
 			cal.add(Calendar.MONTH, 1);
-			cal.add(Calendar.DAY_OF_YEAR, -1); 
+			cal.add(Calendar.DAY_OF_YEAR, -1);
 			m_EndDate = new Timestamp (cal.getTimeInMillis());
 		}
 		log.debug("setStartEndDate = " + p_StartDate + " - " + m_EndDate);
-		
+
 		/**
 		String sd = DB.TO_DATE(p_StartDate, true);
 		StringBuffer sql = new StringBuffer ("SELECT ");
@@ -361,10 +361,10 @@ public class CommissionCalc extends JavaProcess
 				//	CommissionAmount, C_Currency_ID, Amt, Qty,
 				MCommissionDetail cd = new MCommissionDetail (comAmt,
 					rs.getInt(1), rs.getBigDecimal(2), rs.getBigDecimal(3));
-					
+
 				//	C_OrderLine_ID, C_InvoiceLine_ID,
 				cd.setLineIDs(rs.getInt(4), rs.getInt(5));
-				
+
 				//	Reference, Info,
 				String s = rs.getString(6);
 				if (s != null)
@@ -372,11 +372,11 @@ public class CommissionCalc extends JavaProcess
 				s = rs.getString(7);
 				if (s != null)
 					cd.setInfo(s);
-				
+
 				//	Date
 				Timestamp date = rs.getTimestamp(8);
 				cd.setConvertedAmt(date);
-				
+
 				//
 				if (!cd.save())		//	creates memory leak
 					throw new IllegalArgumentException ("CommissionCalc - Detail Not saved");
