+++ de.metas.business/src/main/java-legacy/org/compiere/model/CalloutInOut.java
@@ -29,7 +29,6 @@ import org.adempiere.warehouse.spi.IWarehouseAdvisor;
 import org.compiere.util.DisplayType;
 
 import de.metas.adempiere.form.IClientUI;
-import de.metas.adempiere.model.I_AD_User;
 import de.metas.bpartner.service.BPartnerStats;
 import de.metas.bpartner.service.IBPartnerBL;
 import de.metas.bpartner.service.IBPartnerDAO;
