+++ de.metas.business/src/main/java-legacy/org/compiere/process/ImportOrder.java
@@ -479,7 +479,7 @@ public class ImportOrder extends JavaProcess
 		if (no != 0)
 			log.warn("Invalid Charge=" + no);
 		//
-		
+
 		sql = new StringBuffer ("UPDATE I_Order "
 				  + "SET I_IsImported='E', I_ErrorMsg=I_ErrorMsg||'ERR=Product and Charge, ' "
 				  + "WHERE M_Product_ID IS NOT NULL AND C_Charge_ID IS NOT NULL "
@@ -505,7 +505,7 @@ public class ImportOrder extends JavaProcess
 			log.warn("Invalid Tax=" + no);
 
 		commitEx();
-		
+
 		//	-- New BPartner ---------------------------------------------------
 
 		//	Go through Order Records w/o C_BPartner_ID
@@ -546,9 +546,9 @@ public class ImportOrder extends JavaProcess
 						continue;
 				}
 				imp.setC_BPartner_ID (bp.getC_BPartner_ID ());
-				
+
 				//	BP Location
-				MBPartnerLocation bpl = null; 
+				MBPartnerLocation bpl = null;
 				MBPartnerLocation[] bpls = bp.getLocations(true);
 				for (int i = 0; bpl == null && i < bpls.length; i++)
 				{
@@ -561,8 +561,8 @@ public class ImportOrder extends JavaProcess
 					else if (imp.getC_Location_ID() == 0)
 					{
 						MLocation loc = bpl.getLocation(false);
-						if (loc.equals(imp.getC_Country_ID(), imp.getC_Region_ID(), 
-								imp.getPostal(), "", imp.getCity(), 
+						if (loc.equals(imp.getC_Country_ID(), imp.getC_Region_ID(),
+								imp.getPostal(), "", imp.getCity(),
 								imp.getAddress1(), imp.getAddress2()))
 							bpl = bpls[i];
 					}
@@ -589,18 +589,18 @@ public class ImportOrder extends JavaProcess
 				imp.setC_Location_ID (bpl.getC_Location_ID ());
 				imp.setBillTo_ID (bpl.getC_BPartner_Location_ID ());
 				imp.setC_BPartner_Location_ID (bpl.getC_BPartner_Location_ID ());
-				
+
 				//	User/Contact
-				if (imp.getContactName () != null 
-					|| imp.getEMail () != null 
+				if (imp.getContactName () != null
+					|| imp.getEMail () != null
 					|| imp.getPhone () != null)
 				{
-					List<de.metas.adempiere.model.I_AD_User> users = bp.getContacts(true);
+					List<I_AD_User> users = bp.getContacts(true);
 					I_AD_User user = null;
 					for (int i = 0; user == null && i < users.size();  i++)
 					{
 						String name = users.get(i).getName();
-						if (name.equals(imp.getContactName()) 
+						if (name.equals(imp.getContactName())
 							|| name.equals(imp.getName()))
 						{
 							user = users.get(i);
@@ -638,7 +638,7 @@ public class ImportOrder extends JavaProcess
 			log.warn("No BPartner=" + no);
 
 		commitEx();
-		
+
 		//	-- New Orders -----------------------------------------------------
 
 		int noInsert = 0;
@@ -667,9 +667,9 @@ public class ImportOrder extends JavaProcess
 				if (cmpDocumentNo == null)
 					cmpDocumentNo = "";
 				//	New Order
-				if (oldC_BPartner_ID != imp.getC_BPartner_ID() 
+				if (oldC_BPartner_ID != imp.getC_BPartner_ID()
 					|| oldC_BPartner_Location_ID != imp.getC_BPartner_Location_ID()
-					|| oldBillTo_ID != imp.getBillTo_ID() 
+					|| oldBillTo_ID != imp.getBillTo_ID()
 					|| !oldDocumentNo.equals(cmpDocumentNo))
 				{
 					if (order != null)
@@ -757,10 +757,10 @@ public class ImportOrder extends JavaProcess
 					line.setTax();
 					imp.setC_Tax_ID(line.getC_Tax_ID());
 				}
-				
+
 				// 05129
 				line.setC_TaxCategory_ID(line.getC_Tax().getC_TaxCategory_ID());
-				
+
 				if (imp.getFreightAmt() != null)
 					line.setFreightAmt(imp.getFreightAmt());
 				if (imp.getLineDescription() != null)
@@ -778,7 +778,7 @@ public class ImportOrder extends JavaProcess
 					noInsertLine++;
 					commitEx(); // metas
 				}
-				
+
 			}
 			if (order != null)
 			{
