+++ de.metas.contracts/src/main/java/de/metas/contracts/flatrate/impexp/FlatrateTermImportFinder.java
@@ -1,21 +1,22 @@
 /**
- * 
+ *
  */
 package de.metas.contracts.flatrate.impexp;
+
+import lombok.experimental.UtilityClass;
 
 import java.util.List;
 import java.util.Properties;
 
 import org.adempiere.ad.trx.api.ITrx;
+import org.compiere.model.I_AD_User;
 import org.compiere.model.I_C_BPartner_Location;
 
 import com.google.common.collect.Ordering;
 
-import de.metas.adempiere.model.I_AD_User;
 import de.metas.bpartner.BPartnerId;
 import de.metas.bpartner.service.IBPartnerDAO;
 import de.metas.util.Services;
-import lombok.experimental.UtilityClass;
 
 /*
  * #%L
@@ -27,12 +28,12 @@ import lombok.experimental.UtilityClass;
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
  * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
