+++ de.metas.contracts/src/main/java/de/metas/contracts/subscription/impl/SubscriptionBL.java
@@ -2,6 +2,8 @@ package de.metas.contracts.subscription.impl;
 
 import static org.adempiere.model.InterfaceWrapperHelper.save;
 
+import lombok.NonNull;
+
 /*
  * #%L
  * de.metas.contracts
@@ -60,6 +62,7 @@ import com.google.common.base.Preconditions;
 
 import de.metas.adempiere.model.I_AD_User;
 import de.metas.adempiere.model.I_C_Order;
+import de.metas.bpartner.BPartnerId;
 import de.metas.bpartner.service.IBPartnerOrgBL;
 import de.metas.contracts.Contracts_Constants;
 import de.metas.contracts.FlatrateTermPricing;
@@ -103,7 +106,6 @@ import de.metas.util.Check;
 import de.metas.util.Services;
 import de.metas.util.time.SystemTime;
 import de.metas.workflow.api.IWFExecutionFactory;
-import lombok.NonNull;
 
 public class SubscriptionBL implements ISubscriptionBL
 {
@@ -183,9 +185,9 @@ public class SubscriptionBL implements ISubscriptionBL
 		if (correspondingTerm != null)
 		{
 			correspondingTerm.setC_FlatrateTerm_Next_ID(newTerm.getC_Flatrate_Term_ID());
-			save(correspondingTerm);			
+			save(correspondingTerm);
 		}
-		
+
 		return newTerm;
 	}
 
@@ -390,7 +392,7 @@ public class SubscriptionBL implements ISubscriptionBL
 		newTerm.setBill_Location_ID(bill_Location_ID);
 		newTerm.setBill_User_ID(bill_User_ID);
 
-		newTerm.setDropShip_BPartner_ID(olCandEffectiveValuesBL.getDropShip_BPartner_Effective_ID(olCand));
+		newTerm.setDropShip_BPartner_ID(BPartnerId.toRepoId(olCandEffectiveValuesBL.getDropShip_BPartner_Effective_ID(olCand)));
 		newTerm.setDropShip_Location_ID(olCandEffectiveValuesBL.getDropShip_Location_Effective_ID(olCand));
 		newTerm.setDropShip_User_ID(olCandEffectiveValuesBL.getDropShip_User_Effective_ID(olCand));
 
@@ -932,7 +934,7 @@ public class SubscriptionBL implements ISubscriptionBL
 
 		final ContractOrderService contractOrderService = Adempiere.getBean(ContractOrderService.class);
 		final OrderId orderId = contractOrderService.retrieveLinkedFollowUpContractOrder(currentOrderId);
-		
+
 		if (orderId == null)
 		{
 			return null;
@@ -946,7 +948,7 @@ public class SubscriptionBL implements ISubscriptionBL
 						&& oldTerm.getC_Flatrate_Conditions_ID() == newTerm.getC_Flatrate_Conditions_ID())
 				.findFirst()
 				.orElse(null);
-		
+
 		// check if there is an extended term
 		if (suitableTerm == null)
 		{
@@ -954,10 +956,10 @@ public class SubscriptionBL implements ISubscriptionBL
 		}
 
 		final I_C_Flatrate_Term topTerm = contractOrderService.retrieveTopExtendedTerm(suitableTerm);
-		
+
 		return topTerm == null ? suitableTerm : topTerm;
 	}
-	
+
 
 	@Override
 	public boolean isActiveTerm(@NonNull final I_C_Flatrate_Term term)
