+++ de.metas.ordercandidate.rest-api-impl/src/main/java/de/metas/ordercandidate/rest/JsonConverters.java
@@ -1,5 +1,7 @@
 package de.metas.ordercandidate.rest;
 
+import lombok.NonNull;
+
 import java.util.List;
 
 import org.adempiere.exceptions.AdempiereException;
@@ -22,8 +24,6 @@ import de.metas.pricing.PricingSystemId;
 import de.metas.product.ProductId;
 import de.metas.util.Check;
 import de.metas.util.lang.Percent;
-
-import lombok.NonNull;
 
 /*
  * #%L
@@ -70,6 +70,9 @@ public class JsonConverters
 
 		final CurrencyId currencyId = masterdataProvider.getCurrencyId(request.getCurrencyCode());
 
+		final BPartnerMasterDataProvider //
+		bpartnerMasterdataProvider = masterdataProvider.getBPartnerMasterDataProvider();
+
 		return OLCandCreateRequest.builder()
 				//
 				.orgId(orgId)
@@ -79,10 +82,10 @@ public class JsonConverters
 				//
 				.dataDestInternalName(request.getDataDestInternalName())
 				//
-				.bpartner(masterdataProvider.getCreateBPartnerInfo(request.getBpartner(), orgId))
-				.billBPartner(masterdataProvider.getCreateBPartnerInfo(request.getBillBPartner(), orgId))
-				.dropShipBPartner(masterdataProvider.getCreateBPartnerInfo(request.getDropShipBPartner(), orgId))
-				.handOverBPartner(masterdataProvider.getCreateBPartnerInfo(request.getHandOverBPartner(), orgId))
+				.bpartner(bpartnerMasterdataProvider.getCreateBPartnerInfo(request.getBpartner(), orgId))
+				.billBPartner(bpartnerMasterdataProvider.getCreateBPartnerInfo(request.getBillBPartner(), orgId))
+				.dropShipBPartner(bpartnerMasterdataProvider.getCreateBPartnerInfo(request.getDropShipBPartner(), orgId))
+				.handOverBPartner(bpartnerMasterdataProvider.getCreateBPartnerInfo(request.getHandOverBPartner(), orgId))
 				//
 				.poReference(request.getPoReference())
 				//
@@ -115,14 +118,17 @@ public class JsonConverters
 			return null;
 		}
 
-		final BPartnerId bpartnerId = BPartnerId.ofRepoId(bpartnerInfo.getBpartnerId());
+		final BPartnerId bpartnerId = bpartnerInfo.getBpartnerId();
 		final BPartnerLocationId bpartnerLocationId = BPartnerLocationId.ofRepoIdOrNull(bpartnerId, bpartnerInfo.getBpartnerLocationId());
 		final BPartnerContactId contactId = BPartnerContactId.ofRepoIdOrNull(bpartnerId, bpartnerInfo.getContactId());
 
+		final BPartnerMasterDataProvider //
+		bpartnerMasterdataProvider = masterdataProvider.getBPartnerMasterDataProvider();
+
 		return JsonBPartnerInfo.builder()
-				.bpartner(masterdataProvider.getJsonBPartnerById(bpartnerId))
-				.location(masterdataProvider.getJsonBPartnerLocationById(bpartnerLocationId))
-				.contact(masterdataProvider.getJsonBPartnerContactById(contactId))
+				.bpartner(bpartnerMasterdataProvider.getJsonBPartnerById(bpartnerId))
+				.location(bpartnerMasterdataProvider.getJsonBPartnerLocationById(bpartnerLocationId))
+				.contact(bpartnerMasterdataProvider.getJsonBPartnerContactById(contactId))
 				.build();
 	}
