+++ de.metas.ordercandidate.rest-api/src/main/java/de/metas/ordercandidate/rest/JsonBPartnerInfo.java
@@ -1,10 +1,10 @@
 package de.metas.ordercandidate.rest;
 
-import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility;
-
 import lombok.Builder;
 import lombok.Data;
+
+import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility;
 
 /*
  * #%L
@@ -16,12 +16,12 @@ import lombok.Data;
  * it under the terms of the GNU General Public License as
  * published by the Free Software Foundation, either version 2 of the
  * License, or (at your option) any later version.
- * 
+ *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU General Public
  * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
@@ -33,7 +33,24 @@ import lombok.Data;
 @Builder
 public final class JsonBPartnerInfo
 {
+	public enum IfNotExists
+	{
+		IF_NOT_EXISTS_CREATE,
+
+		IF_NOT_EXISTS_FAIL;
+	}
+
+	public enum IfExists
+	{
+		IF_EXISTS_UPDATE,
+
+		IF_EXISTS_DONT_UPDATE;
+	}
+
 	private JsonBPartner bpartner;
 	private JsonBPartnerLocation location;
 	private JsonBPartnerContact contact;
+
+	private IfNotExists ifNotExists;
+	private IfExists ifExists;
 }
