+++ de.metas.vertical.pharma.msv3.server-peer-metasfresh/src/main/java/de/metas/vertical/pharma/msv3/server/peer/metasfresh/listeners/OrderCreateRequestRabbitMQListener.java
@@ -161,7 +161,7 @@ public class OrderCreateRequestRabbitMQListener
 		}
 
 		return OLCandBPartnerInfo.builder()
-				.bpartnerId(bpartnerId.getBpartnerId())
+				.bpartnerId(de.metas.bpartner.BPartnerId.ofRepoIdOrNull(bpartnerId.getBpartnerId()))
 				.bpartnerLocationId(bpartnerId.getBpartnerLocationId())
 				.build();
 	}
